Extend initial map to 64M
This will allow us to include a big u-root and that u-root can
contain u-root and ALL harvey binaries, i.e. all of amd64/bin.

We're going to want a better compression than just gzip.

-rw-r--r-- 1 rminnich rminnich 70192397 Aug 31 20:43 /tmp/initramfs.plan9_amd64.cpio.gz
-rw-r--r-- 1 rminnich rminnich 23180974 Aug 31 20:25 /tmp/initramfs.plan9_amd64.cpio.lzma

But we are, definitely, going to be able to have pretty much anything we need ... in the initramfs.
This is going to be pretty slick. Boot harvey, get the full graphical environment, THEN mount the
server from 9p. This would include rio etc. This is gonna be very nice.

Graham, tmpfs needs lzma added.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>