centre: tool for supporting harvey DHCP, TFTP, HTTP booting and 9p serving
Centre is intended to be one stop shopping for booting harvey nodes.
It can host DHCP, TFTP, HTTP, and 9p services.
The services are enabled individually.

Centre is configured for IP4 for now, and IP6, soon.

Unlike other systems where config files are scattered all over the place,
Centre can use, e.g., /etc/hosts to hold MAC to IP mappings. However, centre uses
golang host name lookup function, so the host mappings can be anywhere that is used
for hostname to IP mapping.

On Plan 9, for example, the would typically by in /lib/ndb/local, which already
has support for holding MAC and IP assocations.

No special file formats are needed. Users will need the ability to
edit /etc/hosts; but if users are running any kind of pxe or dhcp server,
you've got that already.

Centre need not enable DHCP. At start it looks up the hostname centre and, if
it is found, it will serve DHCP but only for hosts with entries locatable
with net.LookupIP.

When centre receives a request in the dserver4 loop, it will use the
net.LookupIP for a hostname  of the form: 'uMAC', where mac is the stringified
version of net.HardwareAddr, and u is a single-leter prefix to make it a hostname,
not a number. LookupIP returns a []net.IP. The code
uses the 1st element, converted to an ip4 with To4(), as the IP address.

If the lookup fails, centre prints an error message of the form:

2020/07/13 13:55:08 Not responding to DHCP request for mac 00:07:32:4c:25:36
2020/07/13 13:55:08 You can create a host entry of the form 'a.b.c.d [names] u00:07:32:4c:25:36' 'ip6addr [names] u00:07:32:4c:25:36'if you wish

which can be used to create a new hosts file entry.

The lookup does not choose to use /etc/hosts; although the current use is to add mappings to /etc/hosts, they can
be added to any source that is used by net.LookupIP. n

The lookup is stateless; centre does not cache, but rather does a new lookup each time, so that
should the host database, the new information can be used.

Example: one might have a network of several Atomic PIs. One can bring them up one at a time, and look
for messages of the form:
2020/07/13 13:55:08 Not responding to DHCP request for mac 00:07:32:4c:25:36
2020/07/13 13:55:08 You can create a host entry of the form 'a.b.c.d [names] u00:07:32:4c:25:36' 'ip6addr [names] u00:07:32:4c:25:36'if you wish

At that point, one might add this line to /etc/hosts:
192.168.0.4 api3 u00:07:32:4c:04:e6

On the next dhcp request, api3 will be provided with an ip.

This approach has many benefits: the file format is simple; leases need never expire; and, most important, once one assigns
a MAC to IP mapping, it is available for all other programs (e.g. cpu), which has proven very handy in practice.

On a recent experience setting up a small network of atomic pis, the task was accomplished with far less effort, and in
far less time, than was possible with isc-dhcpd and its companion tools.

My current /etc/hosts looks like this:

192.168.0.1 centre
192.168.0.2 up
192.168.0.3 api2 u00:07:32:4b:f9:f3
192.168.0.4 api3 u00:07:32:4c:04:e6
192.168.0.5 api4 u00:07:32:4c:25:36
192.168.0.6 rome uc8:b3:73:1f:44:55
192.168.0.7 a apu2 u00:0d:b9:49:0f:04
192.168.0.8 z zero u00:00:00:00:00:00
192.168.0.239 router

Note that zero was needed for a system that came up with an all-zeros MAC.

centre has options for enabling TFTP, HTTP, and a default host name.

For booting on one Harvey network, we just do this:
centre -i enp2s0f0 -bootfilename harvey.32bit -tftp-dir . -ninepDir ~/harvey

and that is enough. No other directories such as pxelinux.0, or files, or changes
in /etc, are needed.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>