stat for 9P2000.L / devmntn
Sample session with a 9P2000.L server (e.g. ipfs):
srv tcp!192.168.0.1!8080 x
mount -M N /srv/x /n/x
ls -l /n/x

9P2000.L has two big problems: first, a directory read does not return
stat records, as it should; second, the stat has fields missing, particularly
the file name, a string for uid, and a string for gid.

These are hard to work around. The result in output like this:

% ls -l/n/x

------w-r--  0 '' '' 0 Feb  7  2006 /n/x/util

because a readdir does not return stat records, we are missing user, group, size information.

But:
192.168.0.7# ls -ld /n/x
d-rwxrwxr-x N 31 1000 1000 4096 Feb  7  2006 /n/x

Because in this case, the code does a stat on /n/x and gets a stat record back.

Now for the first case, what has to happen, but not in this PR, is on a readdir,
the code must do a full walk and stat *for every directory entry*. This means the
number of RPCs for a directory is 2x the number of directory entries, instead
of 1 total as it is in 9P2000. 64 directory entries will result in 129 RPCs
in 9P2000.L and 1 in 9P2000 (129: 1 for readdir, 128 for 2 RPCs for each entry).

Further, on a stat, the file name is not in the stat; we have to pull the name
from the channel. What happens if the name changed after the initial walk?
You'll have the wrong name in a stat. That's not great.

9P2000.L violates plan 9 rules in other small ways, such as including an inumber
in the stat, because few if any real Plan 9 systems have inumbers, but
Linux had its own requirements.

It does get time right. Plan 9 has a 32-bit time problem; 9P2000.L does not.

It's worth supporting 9P2000.L with all its defects, so here we take
another step on that journey.

The code in devmntn.c:mntstat is nasty. I'd just as soon it stayed that
way, rather than bring too much of 9P2000.L's non-plan9-like nature
into fcall.h and libc.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>