Fix asmalloc if the first element in the list has to be removed
If an asmalloc is large and hence consumes the first element in the list,
asmlist must be bumped to point to the next element. The code, as written,
would end up destroying asmlist and there would be no available memory
on boot. This code worked fine for well over 10 years until we found
an AMD platform with a Reserved region at the 64M mark, as in this
AMD Rome board:

Memory
        0x0000000000001000 0x00000000000a0000 (651264)
Memory
        0x0000000000100000 0x0000000004000000 (66060288)
reserved
        0x0000000004000000 0x0000000004042000 (270336)
Memory
        0x0000000004042000 0x0000000076d00000 (1925963776)

That reserved hole broke asm completely, and the failure was a
platform that bricked. Finding this problem took days, which points
to a deeper problem in Plan 9, which is its horrendous error handling
and lack of a unit test framework.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>