ahci: tidy initialisation sdiahci was hanging when trying to intialise on a Skylake i7-6700K, when calling ahciidle for port 5.  This appears to have been caused by writing to 0x90 in the PCI register.  Commenting it out removes the hang (though I haven't tested beyond that).
There seems to be a bit of init code being called for almost all
Intel controllers - this probably wasn't the original intention.
Given that the PCI registers would change from controller to
controller, this is probably not the best way to do it.  Given
that the code is quite old and therefore the controllers also,
it might be best to remove controller specific init code, follow
the AHCI spec, and see how far we can get.

I restructured the code to put more of the initialisation code in
one place.  This makes it a bit easier to see the same steps being
carried out multiple times.

Signed-off-by: Graham MacDonald <grahamamacdonald@gmail.com>