Add back [DEFG]S registers to Ureg
Commit 6ab2e16c9f1058aab427f79d4516d281c8e6039b removed these
segmentation registers, but we still need them for Go programs to work
correctly.

The Go runtime uses ureg.sp when it needs to print a stack trace after
the program dereferences a nil pointer and receives a note from the
kernel. Since sp comes after the segmentation registers in the struct,
the sp offset within the struct would be wrong if the segmentation
registers are missing.

After this change, Go programs will now panic (and print a stack trace)
instead of crashing when dereferencing a nil pointer.

Fixes #999

Signed-off-by: Fazlul Shahriar <fshahriar@gmail.com>