rename uintmem to physaddr
It was named uintmem because uint is 32-bits and we needed a
physical address type that was kind of uint but really for physical
addresses and hence 64 bits, while, at the same time, not being
tied to a bit length (so u64 would not do).

This all goes back to the days of the 32-bit transition for K10 IIRC.

This is all very confusing so give it a name related to its function.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>