Make ip/ipconfig give up quicker on DHCP
Despite what the man pages say, ip/ipconfig can actually take 300 seconds to time out on acquiring a DHCP lease. That's because it runs up to 30 iterations of a function which itself retries DHCP requests 10 times, once per second. This will instead give up after 20 seconds, which should be more than enough and is much more in-line with what the man page claims.

Signed-off-by: John Floren <john.floren@gravwell.io>