Add support for bringing in Go
if GOLANG_HARVEY is set, the script does a bin mount to sys/go

from there, go will be included.

Tested and "just works".

The amount of memory for qemu had to be bumped to 4096m.

Depending on what is in $GOLANG_HARVEY, the CD may or may not be
big enough.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>