Allow for more Go commands
The format of the command file changes slightly.
The first field is the directory to cd into to build the command.
The rest are the git URL and directory name as before.

Some URLs comtain many commants.
Only commands with a URL are used in the fetch.
Lines like this:
name ...
are not used in a fetch, only a build.

To simplify the build step, use build -o /$objtype/bin/
instead of go install, so even names of the form a/b/c
will work easily.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>