WIP: change rdmsr to be safe
Add labels in rdmsr to indicate which path to take on a failure.

Add GPF handler.

In GPF handler, switch on ureg->pc value, and if an address is handled,
fix uref->pc.

This is basically what is done in linux, minus all the fancy ELF and
linker set machinations. It will work just fine.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>