nboot: get the EFI bootloaders into the right place, add them to usb stick
With this change, on an UP Xtreme i11, we get to the efi bootloader dialog.

From there, type
bootfile=9pccpuf
boot

and watch it reset. We can't handle the UEFI truth yet.

(the UEFI truth is simple: UEFI is really bad, really really bad,
which is why Intel made it our life).

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>